chore: pull clone methods from 'main' branch and fix some assignment oprerators

+++ src/models/RoomExamination.cpp
@@ -58,7 +58,7 @@ void RoomExamination::removeFromWaitingQueue() {
 }
 
 int RoomExamination::getQueueCount() const {
-    return _waitingQueue.size();
+    return static_cast<int>(_waitingQueue.size());
 }
 
 void RoomExamination::acceptWrite(IVisitor* visitor, std::ostream& os) {
